+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Markup attributes may now begin with a digit, letter or underscore. Previously, they were required to begin with a letter or an underscore. This allows the `select` marker to work with numbers: `[select value=1 1=one 2=two 3=three /]`
+
+## [2.2.0] 2022-04-08
+
+### Added
+
 - Added `DeclarationBuilder` and `FunctionTypeBuilder` classes. These classes allow external libraries to construct new `Declaration` and `FunctionType` objects, without having to have access to the internal setters.
 - `CompilationResult.DebugInfo` now provides per-instruction positional debug information.
   - This allows users of the `Compiler` class to access positional information for each instruction, which is an important first step for source-level debugging.